// src/app/interfaces/users.ts
export interface Users {
  address: string
  bankName: string
  bic: string
  city: string
  companyName: string
  country: string
  createdAt: string
  email: string
  firmenbuchnummer: string
  firstName: string
  iban: string
  id: string
  lastName: string
  phone: string
  roles: {
    role: string
  }
  uid: string
  updatedAt: string
  username: string
  zipCode: string,
  password: string
}